refactor: remove dead internal exports
