+++ extensions/onepassword/onepassword-op-path.d.ts
@@ -1,5 +1,3 @@
-export function resolveTrustedExecutablePath(targetPath: string): Promise<string>;
-
 export function resolveTrustedOnePasswordDirectoryPath(targetPath: string): Promise<string>;
 
 export function resolveTrustedOnePasswordCli(options?: {
